+++ src/main/java/de/marhali/easyi18n/util/DocumentUtil.java
@@ -1,0 +1,36 @@
+package de.marhali.easyi18n.util;
+
+import com.intellij.openapi.editor.Document;
+import com.intellij.openapi.fileEditor.FileDocumentManager;
+import com.intellij.openapi.fileTypes.FileType;
+import com.intellij.openapi.vfs.VirtualFile;
+
+public class DocumentUtil {
+    protected Document document;
+    private FileType fileType;
+
+    public DocumentUtil(Document document) {
+        setDocument(document);
+    }
+
+    public Document getDocument() {
+        return document;
+    }
+
+    public void setDocument(Document document) {
+        this.document = document;
+        FileDocumentManager fileDocumentManager = FileDocumentManager.getInstance();
+        VirtualFile virtualFile = fileDocumentManager.getFile(document);
+        if (virtualFile != null) {
+            fileType = virtualFile.getFileType();
+        }
+    }
+
+    public boolean isJsOrTs() {
+        return (fileType.getDefaultExtension().contains("js") || fileType.getDescription().contains("ts"));
+    }
+
+    public boolean isVue() {
+            return fileType.getDefaultExtension().contains("vue");
+    }
+}